Hero component, create highlighted web3 text
